build: convert to node test runner
