+++ types/fs.d.ts
@@ -1,0 +1,4 @@
+export function readFileSync(path: string, encoding: string): string;
+export = {
+  readFileSync
+};
