+++ types/node.test.d.ts
@@ -1,0 +1,3 @@
+declare module 'node:test' {
+  export default function test(name: string, fn: () => unknown | Promise<unknown>): void;
+}
