+++ types/path.d.ts
@@ -1,0 +1,3 @@
+export function join(...paths: string[]): string;
+export const dirname: any;
+export = { join, dirname };
